renepay: add debuging utils
